+++ lib/plausible/stats/comparisons.ex
@@ -99,8 +99,9 @@ defmodule Plausible.Stats.Comparisons do
   defp get_comparison_date_range(source_query, %{mode: "year_over_year"} = options) do
     source_date_range = Query.date_range(source_query, trim_trailing: true)
 
-    start_date = Date.add(source_date_range.first, -365)
-    end_date = source_date_range.last |> Date.add(-365)
+    start_date = source_date_range.first |> Date.shift(year: -1)
+    diff_in_days = Date.diff(source_date_range.last, source_date_range.first)
+    end_date = Date.add(start_date, diff_in_days)
 
     Date.range(start_date, end_date)
     |> maybe_match_day_of_week(source_date_range, options)